Artificial project: limit size of build target population
Summary:
Certain attributes only ever reference a small minority of all
available build targets of a certain type, for example there
is usually a very small number of Java libraries used as an
annotation processor. To reflect this in the generated projects,
keep track of how many unique values we see for each attribute
and attempt to replicate it.

Test Plan:
Run the script, check that the number of targets used as
annotation processors is much smaller than the total number
of Java libraries

Reviewed By: Coneko

fbshipit-source-id: 0ac73ac