Add a target graph transformer to support versioned graphs
Summary:
This is the initial step for version support in Buck.  It adds
a target graph transformer class which takes a normal target graph
containing version-aware nodes and transforms it into a new target
graph with all the versions resolved.

The "special" nodes are:
- `VersionRoot`: the root node (e.g. `cxx_binary`).  Each root defines
  a version sub-graph which will be resolved independently of other
  roots.
- `VersionedAlias`: Node which provides versioning, assigning versions
  to various targets.  When transformed, these nodes will be replaced
  by one of the version choices.
- `VersionPropagator`: The non-root "filler" nodes in a version sub-graph.
  Depending on version resolution, these nodes might get "multiplied" in the
  transformed graph, depending on the different combinations of versions
  selected for their transitive versioned dependencies.

Along with these special nodes, a new "special" parameter, `versionedDeps`,
is supported in constructor args, and are used by version root or propagator
nodes to depend on versioned nodes, along with a constraint used to resolve
the version in the context of the complete version sub-graph.

The strategy for resolving version constraints in a conneced version sub-
graph is configurable, but the only implementation currently added is a
best-effort strategy that runs in O(1).

Currently, just adds a transformer class and several tests.  Subsequent
diffs will add configuration options to enable using this in the UI.

Test Plan: added unittests

Reviewed By: k21

fbshipit-source-id: bade739