macros: add a new `AbstractMacroExpander` base class
Summary:
A very common pattern for `MacroExpander` implementations is to parse
accept macro arguments as strings, parse them into some higher-level
type, then operate on them on in the various interace methods.  This
currently means that the various implementations have a lot of repeated
code, as all their interfaces accept macro arguments as strings which
need to be parsed.

This diff adds a new `AbstractMacroExpander` base class which allows
sub-classes to define a single `parse` method to conver the input args
to a typed input, which is automatically fed through to a new typed
interface.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: d9c01ae