versions: change version universe buck config parsing
Summary:
The previous style of parsing used sections for universes and section
field and values as the target to version mappings.  However, ini
parsing reservers colon for type specifiers and so we can't use build
targets in keys.  This diff moves to makin universes the sections fields,
and specifies their version pinnings as a list of key/values.

Test Plan: added test

Reviewed By: k21

fbshipit-source-id: 1141b11