Remove transitive entries from immediate classpath
Summary:
Running benchmarks on large-scale repositories revealed that we
spend a lot of time in `getTransitiveClasspaths` and in particular
that there is a recursive call
```
getTransitiveClasspaths ->
getImmediateClasspaths ->
getTransitiveClasspaths -> ...
```
This does not seem correct, if I understand the model correctly,
immediate classpaths should not include transitive classpaths of
the library's exported dependencies. Replace
`getTransitiveClasspaths` call with `getImmediateClasspaths` to
fix this and improve performance.

Test Plan: `buck test`

Reviewed By: marcinkosiba

fbshipit-source-id: 256588c