Remove suppliers from BuildRuleParams
Summary:
The fields holding dependencies in `BuildRuleParams` and
`AbstractBuildRule` were modified to be suppliers of a set instead
of a set itself in preparation for a feature that would benefit
from building a small part of the action graph without building
the build rules for all dependencies. The feature was however
never implemented. At the current time, many places seem to be
just passing `Suppliers.instanceOf` and calling `get` on the
returned suppliers, defeating any savings this might have had.
And at least for builds we currently always build the full action
graph before we start computing rule keys, so we don't avoid
building any of the rules anyway.

Test Plan: `buck test`

Reviewed By: Coneko

fbshipit-source-id: 73053e8