pch: don't `-include` original header prior to using the precompiled one
Summary: It's possible that we end up trying to include the original header file, as well as the precompiled version of it.  The original header might end up getting parsed, negating the benefit of using the precompiled version.  Check whether we're trying to use the precompiled version, and if so, don't `-include` the original header directly.

Test Plan:
* local testing shows improved perf
* tests concerning precompiled headers looks good
* watching CI for all other tests

Reviewed By: Coneko

fbshipit-source-id: d706cfc