PipelineNodeCache: don't call the jobSupplier in a synchronized block
Summary:
This removes any contention that may take place in the PipelineNodeCache.
The main purpose of the jobsCache was to make sure only one job is scheduled
for a given key and this can easily be achieved with a ConcurrentHashMap.
This has the nice benefit of not invoking the jobSupplier in the critical
section either.

Test Plan: CI

Reviewed By: k21

fbshipit-source-id: fbdfbc9