Add ReferencecountedWriter.flushAndClose
Summary:
We are still seeing exceptions because we attempt to flush a closed stream. I thought my previous commit would fix that, but it appears that this is still an issue. Given that `commandIdToLogFileHandlerWriter` is a `ConcurrentMap`, and we only ever close old writers obtained with `ConcurrentMap.remove` / `ConcurrentMap.put`, I cannot see how could `GlobalStateManager` close a writer twice. That is under the assumption that `ConcurrentMap` really performs `remove` and `put` atomically. The only explanation I have is that the client closes those writers after obtaining them with `getLogFileHandlerState` which they shouldn't. However, I couldn't find such close calls.
Another possibility is that we have multiple `ReferenceCountedWriters` entries in the `commandIdToLogFileHandlerWriter` map, all which point to the same underlying stream, but without the reference counter being properly updated. I cannot see how would that happen either.

This diff makes the last flush and close be atomic. If we still see this issue it means something is closing the underlying stream without us knowing. If this fixes the issue, it means we don't really understand well the synchronization model in `GlobalStateManager`.

Test Plan: buck logs

Reviewed By: aiked

fbshipit-source-id: 5f42d30