Allow configure IJ SDK by source level
Summary:
Changed `source` option in `java_library` to set `LANGUAGE_LEVEL` in a relative IJ iml file (see integration tests for example);

When `java_library` has `source` option the SDK for the IJ module is determined by an option in config file. For example, for source level 6 (or 1.6) it tries to looks for the following option:

    [intellij]
      java_library_sdk_names = 1.6 => SDK_NAME_1.6, 1.7 => SDK_NAME_1.7

If this option is found, then the value is used as SDK name in the iml file. When there is no value it will use the source level as SDK name, but it uses a unified name (6 will be converted to 1.6). So the logic is almost identical to the old one when there is no such option in `.buckconfig` except unifying source level.

This logic explicitly works for `java_library` only because `IjModuleFactory.getSourceLevel` explicitly checks for `java_library`. It should be pretty simple to enable for `android_library`, but this is out of scope of this change.

Test Plan: Integration tests and local testing.

Reviewed By: marcinkosiba

fbshipit-source-id: 3883fd1