Make the console less spammy when super console is not available
Summary:
When printing to a regular console, only print lines for targets that are built locally.

Printing all the FOUND and CACHE lines can be really spammy when super-console is not available.
Since this is common is scripts that invoke Buck, it is probably better to just print the BUILT lines. This keeps a useful amount of data in long builds, but does not spam the log in simpler cases.
This basically moves these to a higher verbosity level.

Test Plan:
1. Run buck build -v 2 and pipe it into `cat`, see all FOUND & CACHE lines
2. Run buck build with default verbosity, or -v 1, see only lines with BUILT.

Reviewed By: dreiss

fbshipit-source-id: f0d778c