Remove AppendingHasher
Summary:
It looks like some not-so-recent refactoring removed the need
for the `AppendingHasher` class which was originally used by
`RuleKeyBuilder` which needed to call the hasher's `hash()`
method twice. It no longer needs to do this, but we kept using
`AppendindHasher` with 2 internal hashers, which means we
hashed everything that went into the rulekeys twice and threw
away one of the outputs.

Test Plan: CI.

Reviewed By: marcinkosiba

fbshipit-source-id: 35cbb3d