Make swift compiler honor min target sdk version
Summary:
Previously, swift uses version of the sdk it is compiling with as the minimum target version for deployment.

This causes the issue when you compile swift target with newer sdk, the resulting app will crash on the old target event if you specify correct target deployment in buckconfig, for example:

```
[apple]
  iphonesimulator_target_sdk_version = <old_sdk_version>
  iphoneos_target_sdk_version = <old_sdk_version>
```
This PR make swift compiler honor `*_target_sdk_version` settings.
Closes https://github.com/facebook/buck/pull/923

Reviewed By: Coneko

fbshipit-source-id: 3253fb9