Introduce machine-readable logging
Summary:
This introduces machine readable log files for Buck. The concept is that we have MachineReadableLoggerListener and we register events
there if we want them to the logs. The format of saving is one log entry per line with line format `<Event type> <space> <Json of that event>`.
In order to include something in the machine readable log you will need to annotate the event with `JsonView(JsonViews.MachineReadableLog.class)`
so it gets picked up. This is to prevent writing more information that we actually need.

* buckconfig option `machine_readable_logger_enabled` under `log` so it can be disabled.
* updated the docs.
{F63914291}

Test Plan: * runs an integration test that check ParseStarted, ParseFinished events are there and that InvocationInfo object has been serialised and deserialised properly.

Reviewed By: marcinkosiba

fbshipit-source-id: 88411c5