Fix worker process deadlock
Summary:
Fix deadlock that happens when a worker process writes out a lot
of data to stderr. Previously Buck would wait for the process to
write to its stdout and the process would try to write to stderr
and wait for Buck to read it, leading to a deadlock. In the new
code, the process writes its stderr into a file so it does not
have to wait for Buck to read it.

Test Plan: New test that would deadlock before this fix.

Reviewed By: Coneko

fbshipit-source-id: db56909