Use realpath instead of readlink in binary wrappers
Readlink directly returns the target of a symlink. If
the symlink is relative, readlink only returns the
relative path. This path is relative to the buck executable
on $PATH, not to the current working directory, so the
Python invocation fails.

Using realpath instead directly resolves the full absolute
path referred to by the symlink, allowing users to use
relative symlinks in their bin/ directory on $PATH.