Added option to disable use of buck to build when generating project with --focus
Summary:
When using the option --focus, we need to use --build-with-buck together with
it. It's useful because using buck will make the project build properly.
In our use case, we need to generate a small project, and still build it with Xcode.
For it to work, we need to build the static libraries of targets that are not included in --focus before building in Xcode.
We're adding project.xcode_focus_disable_build_with_buck to keep using Xcode to build when generating project with --focus.

Test Plan:
```
$ buck project --focus '//path/to:library //path/to:app' //path/to:app \
    --build-with-buck --config project.xcode_focus_disable_build_with_buck=True
```

Reviewed By: Coneko

fbshipit-source-id: ce17565