pch: reuse PCH rules if same header file + flag combination
Summary:
Previously a PCH was rebuilt for every rule in which the `prefix_header` option was used.  While the resulting compilations were not incorrect, this resulted in much more PCH building than necessary.

This was due to the cache logic for the PCH.  The logic is, compute the hash of target, header, headerlanguage, and compile flags.  This is still the case; however the target part is changed here.  Instead of the target being the one who said `prefix_header`, change it so that if the SourcePath reference in prefix_header parameter is a rule, use the rule's target.

Example which should help un-muddle my poor explanation. :)

Sample BUCK:

  export_file(name='pch', src='header.h')
  cxx_binary(name='bin1', src='bin1.c', prefix_header=':pch')
  cxx_binary(name='bin2', src='bin2.c', prefix_header=':pch')

The header `header.h` would be precompiled twice, once for each binary rule.  That's because the hashes were computed in the two different "instantiations" of the PCH as:

  for bin1: hash(":bin1" + "header.h" + language_C + compile_flags_for_bin1)
  for bin2: hash(":bin2" + "header.h" + language_C + compile_flags_for_bin2)

In both cases everything is the same except the first part, the target, hence different hashes and false cache-misses.

In this new logic, the hash is determined each time to be:

  for bin1: hash(":pch" + "header.h" + language_C + compile_flags_for_bin1)
  for bin2: hash(":pch" + "header.h" + language_C + compile_flags_for_bin2)

So if compile flags are the same for bin1 and bin2, then the PCH is reused.

In addition to the unit tests I tried invoking PCH for a substantial-sized test build.  I saw 722 usages of the PCH (i.e. 722 different .cpp's were built), and within that there are 313 unique combinations (in the previous less-optimized case).  Even in those unique combinations there are repeated build flags, headers, etc.  In terms of wall-time, using PCH vs. without was almost equal, the PCH case was a bit higher.

In the new case there are only 208 PCH's built this time, which seems better, but still high.  Wall time was unaffected.

Test Plan:
* `ant clean ; ant` built ok.
* `ant -Dtest.class=*Precompi*Test java-test` works; note the addition of some test cases in this diff.  The new cases were breaking previously and they now work with the new logic.  Existing test cases pass unscathed.
* java linter is happy
* watching CI for the remainder of test results.

Tested a build locally, went ok.

  BUCK=~/devtools/buck/bin/buck
  TESTBUILD='//path/to/some/project:example_unit_test'
  USE_PREFIX_HEADER='//tools/build/pch:stdcxx'
  NO_BUCKD=1
  export USE_PREFIX_HEADER NO_BUCKD
  killall -9 java ; sleep 1 ; rm -rf buck-out/
  /usr/bin/time $BUCK build mode/dev --no-cache -j 20 -c cxx.untracked_headers=ignore $TESTBUILD

Reviewed By: yiding

fbshipit-source-id: 3e0bd66