Use a proper topo-sort for merged native libraries
Summary:
We need to topo-sort the merged constituents objects to ensure we
processes them in bottom-up order.  Previously, I was using
AcyclicDepthFirstPostOrderTraversal, with some manual processing to
detect the roots.  However, that processing was incorrect when a cycle
with no incoming edges was present.  Now, use TopologicalSort instead.

This actually results in much smaller code, but I had to add a lot of it
back to produce a clear error message when a cycle is found.  I feel
like that is still a significant complexity win because the
error-message-generating code isn't as important.

Test Plan: Added an integration test that failed with the old case.

Reviewed By: asp2insp

fbshipit-source-id: 2ddcefe