Eliminate the pause after a failed build
Summary:
... by defaulting the `MissingSymbolsHandler` code to off.

Note that this is a 'conservative' solution that retains the buggy (?) code, in case someone ever wants to debug/extend the missing-symbol callback which (arguably) //is// a lot cleaner / more elegant than parsing `stderr`.

Test Plan: Introduce a  deliberate error, like zapping a dep in a `BUCK` file. Compilation stops right after the suggestions, instead of sitting silently for a variable amount of time (sometimes it's only ten or fifteen seconds; I've seen it push two minutes).

Reviewed By: asp2insp

fbshipit-source-id: 770574c