Don't check ignored paths in DefaultFileHashCache (by default)
Summary:
This check is taking up a considerable amount of time in no-op builds,
and I've never heard of it catching an actual issue.  Let's disable it
by default.  We can set a system property to re-enable it in CI to check
for bugs.

Test Plan: Saw no-op builds being faster and more consistent.

Reviewed By: JonShemitz

fbshipit-source-id: 45a5b93