Handle ExecutionExceptions more consistently
Summary:
If an `ExecutionException` contains a checked exception, rethrow
it and handle it as if it was thrown directly. If it contains an
unchecked exception, rethrow it as a `RuntimeException` (as
opposed to just posting a console event).

Test Plan:
Add code that throws an unchecked exception into
`CachingBuildEngine`, check that the exception is propagated.

Reviewed By: Coneko

fbshipit-source-id: 53592ef