Make application test work when xctest is forced on
Summary:
Currently when `useXctest` is forced on, we lose the test host application information.
This is causing us not able to run the application test properly when running `buck test`
when a platform is configured to always run with xctest.
In this patch, we pass the test host application path to environment `XCInjectBundleInto`
when invoking `xctest`, which will cause the application to become the host app.
Note `XCInjectBundleInto` is ignored by lower versions of `xctest`.

Test Plan:
1) Force the use of xctest by adding the testing platform to `xctest_platforms` in `.buckconfig`.
2) Config a 'apple_test' with desired `test_host_app` set.
3) Run the test with `buck test`, make sure the test host app would start and tests would run.

Reviewed By: Coneko

fbshipit-source-id: ae9bc93