Chase down the ReferenceCountedWriter stream closed issue
Summary:
Okay, this fixes another race condition and I hope this is it now :)
What could've happend before is that defaultWriter was put in the map, then before a new reference has been created out of it, another thread kicks in, removes it from the map and finally closes it. Only then the original thread creates a new reference, but out of an already closed writer. This was not caught because there was no check. I added a check and also refactored the code to avoid this issue.

Test Plan: buck tests

Reviewed By: Coneko

fbshipit-source-id: 04a3050