Introduce include_signature option under [autodeps] in .buckconfig.
Summary:
Historically, `BUCK.autodeps` files have contained a line like:

`#@# GENERATED FILE: DO NOT MODIFY c9a06044ccac3d1356249ab06a01f82874510b5d #@#`

to use to insert a checksum to help prevent developers from editing the contents
of a `BUCK.autodeps` file by hand (as opposed to running `buck autodeps`).

In practice, this has been annoying to deal with when performing merges,
so this introduces an option to disable the signatures. Note that the signatures
will still be enabled by default.

Historically, the dirty secret has been that the signatures were not actually verified
(see how `VERIFY_AUTODEPS_SIGNATURE = False` is removed in `buck.py` in this revision)
because we weren't sure how well this whole autodeps thing was going to work out,
so we wanted to disable signature verification during this trial period in case we
needed to make a quick fix.

Now that you can opt out of autodeps signatures via in `.buckconfig`, we will always verify
a signature if a `BUCK.autodeps` file includes one. (We should probably pass the setting
of `include_signature` into `buck.py`, but I wanted to keep things simple, for now.)

Test Plan:
In a follow-up revision, re-run `buck autodeps` on Buck itself with the new setting
and verify that it still builds.
Existing tests ensure that things still work when signatures are included.
In particular, now that `VERIFY_AUTODEPS_SIGNATURE = False` has been removed,
the existing tests are even stronger because the signatures are now verified
when `BUCK.autodeps` is read.

Reviewed By: marcinkosiba

fbshipit-source-id: 9a43be9