Enable autodeps signature verification in Buck.
Summary:
Historically, the dirty secret has been that autodeps signatures were not
actually verified (see how `VERIFY_AUTODEPS_SIGNATURE = False` is set to `True`
in `buck.py` in this revision). Initially, we weren't sure how well this whole
autodeps thing was going to work out, so we wanted to disable signature
verification during this trial period in case we needed to make a quick fix.

It turns out that the logic for verifying signatures in `BUCK.autodeps` files
was never tested with the output of `AutodepsWriter`. The changes to `buck.py`
were introduced first (before `AutodepsWriter`) in D2877122. In that revision,
the `BUCK.autodeps` files used in the integration tests were written "by hand"
in the sense that they were written without the signature line, run through
`shasum`, and then the header was added manually. Indeed, if you check out that
revision, modify `buck.py` to be:

```
VERIFY_AUTODEPS_SIGNATURE = True
```

and then run the test that was included as part of that revision, it passes:

```
ant java-test -Dtest.class=ParserIntegrationTest
```

In that sense, D2877122 was internally consistent even though signature checking
was disabled.

However, it was D2909170 that introduced `AutodepsWriter`. The changes
introduced in that revision were not tested with `VERIFY_AUTODEPS_SIGNATURE` set
to `True`. If they were, we would have discovered the inconsistency where:

* `AutodepsWriter` excludes the trailing newline when calculating the signature.
* `buck.py` includes the trailing newline when calculating the signature.

Because the signature is already a function of pretty-printed JSON, whitespace
is already significant, so for consistency, it seems that the trailing newline
should also be included in the signature. This means that we must change the
behavior of `AutodepsWriter` rather than `buck.py` in order to fix everything,
which is why all of the `BUCK.autodeps` files must be updated as part of this
revision. (Also, the comment in `AutodepsWriter`, as it was originally written, was
mistaken. Where it said "included in," it should have said, "excluded from."
This comment is rewritten as part of this revision since the logic there is also
modified.)

Finally, the signatures of the `BUCK.autodeps.expected` that were introduced as
part of D2909170 (which introduced `AutodepsWriter`) must also be updated as
part of this revision.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 089db7d