Improve heuristics for `buck autodeps`.
Summary:
When `org.stringtemplate.v4.ST` is used as a generic in a public method param,
it should be treated as an exported dependency rather than assumed to be a
template parameter even though it looks like one given its casing.

This fixes an issue that we had to workaround in D4009953.

Test Plan: Added a test case in `JavaFileParserTest` that reproduces the issue.

Reviewed By: mzlee

fbshipit-source-id: 0749dcf