Add exported_deps param to standard_java_library() for autodeps workaround.
Summary:
Reverts change in D4009953 that replaced a `standard_java_library()` with a
`java_library()` by adding support for `exported_deps` in
`standard_java_library()` so that it's no longer necessary to throw it out.

I have an upcoming revision that should eliminate the need for the use of the
explicit use of `exported_deps` here, as that is due to a bug in `buck autodeps`.

Test Plan: Ran `buck autodeps` and made sure everything still builds.

Reviewed By: mzlee

fbshipit-source-id: 56cac48