Use include_signature option to make interpretation of BUCK.autodeps files stricter.
Summary:
Previously, `buck.py` would peek at the first line of a `BUCK.autodeps` file:
* If it matched the signature regex, it would assert the signature matched.
* If it did not match the signature regex, it would parse the entire file as JSON.

This is a little weird in that instead of "peeking", Buck should know which file
format to expect based on the `include_signature` setting in `.buckconfig`. This
updates Buck to enforce this setting when parsing `.autodeps` files.

Test Plan: Make sure Buck still builds with this additional strictness.

Reviewed By: marcinkosiba

fbshipit-source-id: 94ae94d