Thread clockId through to the WatchmanWatcher
Summary:
Add a `project.watchman_cursor` configuration to control whether we use a clockId or a named cursor for the WatchmanWatcher query.

The watchmanQueryUUID is made optional, but created unless `project.watchman_cursor` is configured to `clock_id` and watchman is recent enough (v4.7) to provide a clock command.

Test Plan: Update the unit and integration tests

Reviewed By: marcinkosiba

fbshipit-source-id: 6243aed