Pass the execution context to the step runner when running steps.
Summary:
The step runner only needs the execution context when running steps,
not otherwise. Change it not to be a constructor arg, so that we can
change the execution context to be scoped per execution rather than
per build. This allows the step runner to be created more easily, so
we can remove it from the build context.

Test Plan: CI

Reviewed By: plamenko

fbshipit-source-id: 732cea5