Add option to configure parser Python interpreter
Summary:
The parser has been re-using the parser.interpreter setting, but this is meant
for build rules; the parser needs to have its own independent settings.

This is done in a new class rather than in the ParserConfig view on
ConfigParser, as that has no access to an ExecutableFinder.

Test Plan: Run `ant test` and `buck test`.

Reviewed By: Coneko

fbshipit-source-id: 8e4e875