Make DirArtifactCache be more reasonable in absence of atime
Summary:
It isn't uncommon for a filesystem to not have atime for files.
Currently, if that's the case, the DirArtifactCache eviction makes no
sense (as all files will return the same default atime value). This
change makes the cache fall back to evicting based on ctime in that
case.

Test Plan: added test

Reviewed By: marcinkosiba

fbshipit-source-id: 2a0766a