Always output pexes to unflavoured location.
Summary:
We can do this since the only case when the location will be flavoured
is when it's the binary of a python_test, and it will not clash with
that. This allows us to keep the path of pexes short, which would
otherwise prevent them from running correctly on Windows.

Test Plan: CI

Reviewed By: andrewjcg

fbshipit-source-id: 1fb6483