Distributed build now re-hashes all external links when building
Summary:
This change ensures that all files outside of the project, pointed to by symlinks, are re-hashed when the actual build happens. We can then be sure that those files are identical to the ones on the client that started the distributed build.

If the hashes do not match, or files are missing when the build happens, an error will be thrown.

Test Plan: ci

Reviewed By: marcinkosiba

fbshipit-source-id: ab9fcbd