introduce logic to buck doctor
Summary:
The new buck doctor command is going to be used to help users debug their build by
providing fixes or suggestions. The command gathers information about only **one** build
and sends it to a remote endpoint that is doing the debugging and returns the suggested
solutions.

The communication is happening in Json or in a post request.
If more arguments are needed then an extra section needs to be defined in the form of
[doctor_extra_request_args] and all the fields are added as parameters in the request.

Both are supported and the values can be set in the buckconfig along with the endpoint.

Format of request: `AbstractDoctorEndpointRequest`.
Format of response: `AbstractDoctorEndpointResponse`

Documentation:
{F64093755}
{F64093756}

Note: the documentation is missing the rage stuff, it will be done in another diff since it's unrelated to this.

Test Plan:
- new integration test that constructs, communicates and present the information.
- tests for error messages

Reviewed By: marcinkosiba

fbshipit-source-id: 80dd631