InputBasedRuleKeyBuilderFactory: make size limit per-rule
Summary:
Previously, the change to add input-based rule key size limits made
the limits members of the factory, which meant the factory would
stop producing input-based rule keys after calculating enough rule
keys to push it past the limit.  This also meant that whether we'd
get an input-based rule key for a particular rule would be non-
deterministic, as it'd depend on the exact build targts included in
the build and the order by which they were built.

This diff makes the size limits members of the per-rule builder class,
so that the size limit calculation is done per-rule.

Test Plan: added unittest

Reviewed By: k21

fbshipit-source-id: c1b61af