Remove an unnecessary check for LinearAlloc in debug builds
Summary:
This check was added in the early days of pre-dexing on the assumption
that it would be a very rare error.  However, in a world where gradle
users are converting large apps to Buck, it's been happening more often
and causing unnecessary pain.  I think we can drop this because:
- Most of our developers do not target Gingerbread.
- Those that do don't necessarily need their *debug* builds to work on
  Gingerbread (this check doesn't apply to release builds).
- Our ability to accurately predict LinearAlloc exhaustion on this
  codepath is actually quite weak.

Test Plan: Gautam from Uber has been trying this locally with success.

Reviewed By: asp2insp

fbshipit-source-id: 51a3daa