Rename LinearAllocEstimate to "weight" in dex merging
Summary:
Some parts of buck actually deal with LinearAlloc footprint estimates,
but many of them use the term to refer to an abstract (and very coarse)
estimate of how close a merged dex file is to hitting some sort of
limit (LinearAlloc, method limit, etc.).  Rename the latter to "weight",
which properly conveys that this is an abstract estimate.  All renaming
references to "LinearAlloc" now *actually* mean LinearAlloc.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 01137c9