Always use_linear_alloc_split_dex in android_binary
Summary:
Because zip splitting is no longer done for debug builds, it no longer
needs to be super fast.  Therefore, we no longer need to support the
LinearAlloc-oblivious zip splitter.  Drop it and simplify a bunch of
code.

Add a warning if the argument is present, but we can drop this after a
while.  It was never documented.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: cdbd7d7