Merge certain entitlement keys that should be included
Summary:
Some entitlements were incorrectly excluded from being merged when signing a iOS bundle.

This would manifest itself in the edge case where we added one of these entitlements to an app, but Buck still chose an old provisioning profile without the newly-added entitlements.

This caused Buck-built `.ipa`s to be uninstallable on devices (or not able to be submitted to the App Store) if their profile didn't have the right entitlements.

Test Plan: Build apps with those entitlements and submit them to App Store.

Reviewed By: dinhviethoa

fbshipit-source-id: 5444f1d