Prefer Sha1HashCode.update(Hasher) to Hasher.putBytes(Sha1HashCode.getHash()).
Summary:
`Sha1HashCode.update(Hasher)` is the preferred way of doing this because it is cheaper.
`getHash()` should be reserved for debugging/unit tests.

This also replaces instances of `Sha1HashCode.toString()` with `Sha1HashCode.getHash()`
so the use of the method is explicit and therefore easier to audit.

We could go so far as to move `getHash()` to a package-private test utility to
help prevent this mistake, but that would make `Sha1HashCode.toString()` tricky
to implement. (Also, there are some legitimate uses of `getHash()` in `src/`,
so this really isn't feasible.)

D4074507 prompted this investigation.

Test Plan: CI

Reviewed By: plamenko

fbshipit-source-id: 27b4689