Allow all build rules to experiment with autodeps.
Test Plan:
Added integration test that verifies autodeps can be specified for a `python_library()` rule.
Also added integration test that verifies that when `autodeps=True` is added for a
`python_library()` rule, `buck autodeps` does not complain even though it does not know how
to generate dependencies for such a rule.

Reviewed By: marcinkosiba

fbshipit-source-id: be98558