pch: don't `-include` original header + its PCH version
Summary: When using a `prefix_header`, the header is added to the build with `-include headerfilename.h` -- even if the precompiled version is also used with `-include-pch headerfilename.h.pch`.  This fixes the command-line building so that Buck will now use the `-include` option for a prefix header if and only if the PCH is not used.

Test Plan:
* CI for existing tests
* passes new tests added here

Reviewed By: Coneko

fbshipit-source-id: ae6ecce