Move Java symbol extraction logic so it can be used as a standalone tool.
Summary:
Basically, all this does is move the business logic that was previously in an
instance method of `JavaLibrarySymbolsFinder` and turns it into a static
utility method in the new `SymbolExtractor` class.

It then creates a small `java_library()` for `SymbolExtractor.java` so we can create
a command-line tool for the symbol extractor in an upcoming revision.
You can verify that the `deps` of the new
`//src/com/facebook/buck/jvm/java/autodeps:SymbolExtractor` are pretty slim,
as desired:

```
  "SymbolExtractor" : {
    "deps" : [ ],
    "exported_deps" : [
      "//src/com/facebook/buck/jvm/java:autodeps",
      "//third-party/java/guava:guava",
      "//third-party/java/jackson:jackson-databind"
    ]
  },
```

However, `//src/com/facebook/buck/jvm/java:autodeps` pulls in quite a bit of stuff,
so I may have to do some cleanup if I want to end up with a `java_binary()` for
`SymbolExtractor` that is slim.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 9db6c51