parser: install builtins via a `with`-able function
Summary:
The current way to install builtins for the pyton parser has the
side-effect of changing the contents of the `__builtin__` module.
This can cause some of the tests to fail in strange, hard-to-diagnose
ways if you forget to re-install the builtins (as it'll start using
the state from previous test if not).  This diff switches to expose
the install function via a method meant to be used with `with` which
restores the original namespace and makes the process less side-effect-y.

Test Plan: CI

Reviewed By: aiked

fbshipit-source-id: 7eed18b