disable VCS info gathering from most commands
Summary:
This adds an extra field on each command that shows if we want to gather source control information
while executing this command. This is because gathering the information is slowing down some
commands and does not necessarily provide some useful information.

You can still force the gathering of the information by adding the following to buckconfig:

  [version_control]
      generate_statistics = true

Test Plan: Checked that invoking commands that don't gather the data actually don't gather the data.

Reviewed By: marcinkosiba

fbshipit-source-id: df86325