move MoreStreams.copyExactly to ThriftArtifactCacheProtocol
Summary:
There is only one consumer of this, and this name would be much more sensible
for additional java 8 stream related utilities, instead of byte stream related
utilities.

Also rewrite this in terms of ByteStreams functions, which makes it seem much
less complicated that it was originally.

Test Plan: ci

Reviewed By: Coneko

fbshipit-source-id: da753ed