fix baseRevision for Rage's VcsInfoCollector
Summary:
While making changing in the way we collect the VCS info I made a mistake and
made rage pick up changed files and the baseRevisionId like when we execute
normal commands. This is wrong because in rage we want to get the diff between
the current status and the base revision. This diff addresses this issue.

Test Plan:
* CI
* manual testing of buck rage

Reviewed By: Coneko

fbshipit-source-id: 9df84a6