Stop passing around WatchmanDiagnosticCache by folding its functionality into a BuckEventBus listener
Summary:
I got tired of seeing WatchmanDiagnosticCache passed around. Since its sole
function is to deduplicate diagnostic messages that's going over the eventbus, I
thought it would be fitting to simply have it handle a special wrapped message,
which gets unwrapped and re-posted to the event bus if it's not a duplicate.

Test Plan: CI.

Reviewed By: Coneko, bhamiltoncx

fbshipit-source-id: 449dcef