Return diagnostics as list, not set
Summary:
I sent this out a while ago, but the diff bit-rotted. This
restores the diff on master.

We'll want this to ensure diagnostics returned by `buck.py` are
ordered deterministically and that fatals actually fatal the request
before other diagnostics are processed.

Test Plan:
New test added to ensure order.

  ./bin/buck test //src/com/facebook/buck/json:buck_test

Reviewed By: Coneko

fbshipit-source-id: ee4c197