Revert "Actually format log messages"
Summary:
This reverts commit 0fe1f86ce3b242e7b7ba2b24e78fdcf410753e0c.

The assumption behind this commit was that we were mixing `{0}` and
`printf` style formatting and failing to format the latter.

This wasn't actually true, and this change introduced a bug
where any message containing a `%` post-format started raising
errors.

I will follow this up with a test to confirm the desired behavior.

Test Plan: revert-hammer

Reviewed By: illicitonion

fbshipit-source-id: 6e65c9c